+++ run_tests.py
@@ -13,27 +13,27 @@ from tests.VacancyTest import VacancyTest
 from tests.ExecutorRegTest import ExecutorRegTest
 from tests.ClientRegTest import ClientRegTest
 from tests.ProfileTest import ProfileTest
-from tests.MainTest import MainTest
+from tests.MainPageTest import MainPageTest
 from tests.NavbarsTest import NavbarsTest
 from tests.SettingsTests import SettingsTests
 
 if __name__ == '__main__':
     suite = unittest.TestSuite((
-        # unittest.makeSuite(NavbarsTest),
+        unittest.makeSuite(NavbarsTest),
         unittest.makeSuite(AuthTest),
         unittest.makeSuite(SearchTest),
         unittest.makeSuite(AllOrdersTest),
-        # unittest.makeSuite(CreateOrderTest),
-        # unittest.makeSuite(OrderTest),
-        # unittest.makeSuite(ChangeOrderTest),
-        # unittest.makeSuite(CreateVacancyTest),
-        # unittest.makeSuite(VacancyTest),
-        # unittest.makeSuite(ChangeVacancyrTest),
+        unittest.makeSuite(CreateOrderTest),
+        unittest.makeSuite(OrderTest),
+        unittest.makeSuite(ChangeOrderTest),
+        unittest.makeSuite(CreateVacancyTest),
+        unittest.makeSuite(VacancyTest),
+        unittest.makeSuite(ChangeVacancyrTest),
         unittest.makeSuite(ExecutorRegTest),
         unittest.makeSuite(ClientRegTest),
         unittest.makeSuite(ProfileTest),
-        # unittest.makeSuite(MainTest),
-        # unittest.makeSuite(SettingsTests),
+        unittest.makeSuite(MainPageTest),
+        unittest.makeSuite(SettingsTests),
     ))
 
     result = unittest.TextTestRunner().run(suite)
